pyproject.toml
[project]
name = "mcp-openrouter-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "mcp[cli]>=1.9.4",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
